requirements.txt
# Main training requirements
tensorflow == 1.15.0
numpy == 1.18.1
progressbar2
pandas
six
pyxdg
attrdict
absl-py

# Requirements for building native_client files
setuptools

# Requirements for transcribe.py
webrtcvad

# Requirements for importers
sox
bs4
requests
librosa
soundfile

# Miscellaneous scripts
paramiko >= 2.1
scipy
matplotlib
opuslib >= 3.0.1
simpleaudio
